Fix len_base funcs

+++ ft_uintlen_base.c
@@ -6,7 +6,7 @@
 /*   By: dpoveda- <me@izenynn.com>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/27 14:53:56 by dpoveda-          #+#    #+#             */
-/*   Updated: 2021/09/27 14:53:57 by dpoveda-         ###   ########.fr       */
+/*   Updated: 2021/09/27 15:03:23 by dpoveda-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@ int	ft_uintlen_base(unsigned int n, int base_len)
 	int	cnt;
 
 	cnt = 1;
-	while (n >= base_len)
+	while (n >= (unsigned int)base_len)
 	{
-		n /= base_len;
+		n /= (unsigned int)base_len;
 		cnt++;
 	}
 	return (cnt);
