+++ ft_ulonglen_base.c
@@ -6,7 +6,7 @@
 /*   By: dpoveda- <me@izenynn.com>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/27 14:54:06 by dpoveda-          #+#    #+#             */
-/*   Updated: 2021/09/27 14:54:10 by dpoveda-         ###   ########.fr       */
+/*   Updated: 2021/09/27 15:03:08 by dpoveda-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ int	ft_ulonglen_base(unsigned long n, int base_len)
 	int	cnt;
 
 	cnt = 1;
-	while (n >= base_len)
+	while (n >= (unsigned long)base_len)
 	{
 		n /= base_len;
 		cnt++;
